reorganize config; merge home into system
it's good this way, trust me (im clueless when it comes to nix).
the idea is that it's more important to separate concerns by what *feature* some config provides on a system, rather than where the config files go.
sometimes one feature requires coupled config across system and home files, so the conventional separation of system and home files is actually not good.
that convention makes it harder to group configs along the important axis, and doesnt offer much thats useful in its stead.